+++ changelog.markdown
@@ -160,4 +160,6 @@ and added the functions; `Factory.Data.Interval.precisely`, `Factory.Data.Interv
 ## 0.3.1.0
 * Replaced  '+negate' with 'subtract'.
 * Removed **Eq** type-constraint from **Factory.Math.Statistics.getWeightedMean**.
-* Addes specialisation-pragmas to **Factory.Math.Statistics**.
+* Added specialisation-pragmas to **Factory.Math.Statistics**.
+## 0.3.1.1
+